+++ CHANGELOG.md
@@ -4,7 +4,11 @@ Changelog for paco.models
 6.3.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+### Added
+
+- IRoute53HealthCheck has ip_address field.
+
+- resource/snstopics.yaml is an alias for resource/notificationgroups.yaml
 
 
 6.3.0 (2019-12-03)